requirements.txt
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cloudinary==1.36.0
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
firebase==4.0.1
firebase-admin==6.2.0
Flask==2.2.5
flask-restx==1.1.0
Flask-Mail==0.9.1
gcloud==0.17.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
gunicorn==21.2.0
jsonschema-specifications== 0.1.3
jws==0.1.3
MarkupSafe==2.1.3
msgpack== 1.0.5
oauth2client==3.0.0
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase
python-jwt==2.0.1
pytz==2023.3.post1
referencing==0.8.9
requests
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.6
Werkzeug== 2.2.3
